git-svn-id: https://svn.apache.org/repos/asf/hive/branches/tez@1622108 13f79535-47bb-0310-9956-ffa450edef68

+++ ql/src/java/org/apache/hadoop/hive/ql/exec/Utilities.java
@@ -111,6 +111,7 @@ import org.apache.hadoop.hive.ql.exec.mr.ExecDriver;
 import org.apache.hadoop.hive.ql.exec.mr.ExecMapper;
 import org.apache.hadoop.hive.ql.exec.mr.ExecReducer;
 import org.apache.hadoop.hive.ql.exec.mr.MapRedTask;
+import org.apache.hadoop.hive.ql.exec.tez.DagUtils;
 import org.apache.hadoop.hive.ql.exec.tez.TezTask;
 import org.apache.hadoop.hive.ql.io.ContentSummaryInputFormat;
 import org.apache.hadoop.hive.ql.io.HiveFileFormatUtils;
@@ -1362,8 +1363,8 @@ public final class Utilities {
       codecClass = FileOutputFormat.getOutputCompressorClass(jc, DefaultCodec.class);
       codec = (CompressionCodec) ReflectionUtils.newInstance(codecClass, jc);
     }
-    return (SequenceFile.createWriter(fs, jc, file, keyClass, valClass, compressionType, codec,
-	progressable));
+    return SequenceFile.createWriter(fs, jc, file, keyClass, valClass, compressionType, codec,
+      progressable);
 
   }
 
@@ -1536,6 +1537,24 @@ public final class Utilities {
    */
   private static final Pattern FILE_NAME_TO_TASK_ID_REGEX =
       Pattern.compile("^.*?([0-9]+)(_[0-9]{1,6})?(\\..*)?$");
+
+  /**
+   * Some jobs like "INSERT INTO" jobs create copies of files like 0000001_0_copy_2.
+   * For such files,
+   * Group 1: 00000001 [taskId]
+   * Group 3: 0        [task attempId]
+   * Group 4: _copy_2  [copy suffix]
+   * Group 6: copy     [copy keyword]
+   * Group 8: 2        [copy file index]
+   */
+  private static final Pattern COPY_FILE_NAME_TO_TASK_ID_REGEX =
+      Pattern.compile("^.*?"+ // any prefix
+                      "([0-9]+)"+ // taskId
+                      "(_)"+ // separator
+                      "([0-9]{1,6})?"+ // attemptId (limited to 6 digits)
+                      "((_)(\\Bcopy\\B)(_)"+ // copy keyword
+                      "([0-9]{1,6})$)?"+ // copy file index
+                      "(\\..*)?$"); // any suffix/file extension
 
   /**
    * This retruns prefix part + taskID for bucket join for partitioned table
@@ -1862,26 +1881,70 @@ public final class Utilities {
           // speculative runs), but the largest should be the correct one since the result
           // of a successful run should never be smaller than a failed/speculative run.
           FileStatus toDelete = null;
-          if (otherFile.getLen() >= one.getLen()) {
-            toDelete = one;
+
+          // "LOAD .. INTO" and "INSERT INTO" commands will generate files with
+          // "_copy_x" suffix. These files are usually read by map tasks and the
+          // task output gets written to some tmp path. The output file names will
+          // be of format taskId_attemptId. The usual path for all these tasks is
+          // srcPath -> taskTmpPath -> tmpPath -> finalPath.
+          // But, MergeFileTask can move files directly from src path to final path
+          // without copying it to tmp path. In such cases, different files with
+          // "_copy_x" suffix will be identified as duplicates (change in value
+          // of x is wrongly identified as attempt id) and will be deleted.
+          // To avoid that we will ignore files with "_copy_x" suffix from duplicate
+          // elimination.
+          if (!isCopyFile(one.getPath().getName())) {
+            if (otherFile.getLen() >= one.getLen()) {
+              toDelete = one;
+            } else {
+              toDelete = otherFile;
+              taskIdToFile.put(taskId, one);
+            }
+            long len1 = toDelete.getLen();
+            long len2 = taskIdToFile.get(taskId).getLen();
+            if (!fs.delete(toDelete.getPath(), true)) {
+              throw new IOException(
+                  "Unable to delete duplicate file: " + toDelete.getPath()
+                      + ". Existing file: " +
+                      taskIdToFile.get(taskId).getPath());
+            } else {
+              LOG.warn("Duplicate taskid file removed: " + toDelete.getPath() +
+                  " with length "
+                  + len1 + ". Existing file: " +
+                  taskIdToFile.get(taskId).getPath() + " with length "
+                  + len2);
+            }
           } else {
-            toDelete = otherFile;
-            taskIdToFile.put(taskId, one);
-          }
-          long len1 = toDelete.getLen();
-          long len2 = taskIdToFile.get(taskId).getLen();
-          if (!fs.delete(toDelete.getPath(), true)) {
-            throw new IOException("Unable to delete duplicate file: " + toDelete.getPath()
-                + ". Existing file: " + taskIdToFile.get(taskId).getPath());
-          } else {
-            LOG.warn("Duplicate taskid file removed: " + toDelete.getPath() + " with length "
-                + len1 + ". Existing file: " + taskIdToFile.get(taskId).getPath() + " with length "
-                + len2);
+            LOG.info(one.getPath() + " file identified as duplicate. This file is" +
+                " not deleted as it has copySuffix.");
           }
         }
       }
     }
     return taskIdToFile;
+  }
+
+  public static boolean isCopyFile(String filename) {
+    String taskId = filename;
+    String copyFileSuffix = null;
+    int dirEnd = filename.lastIndexOf(Path.SEPARATOR);
+    if (dirEnd != -1) {
+      taskId = filename.substring(dirEnd + 1);
+    }
+    Matcher m = COPY_FILE_NAME_TO_TASK_ID_REGEX.matcher(taskId);
+    if (!m.matches()) {
+      LOG.warn("Unable to verify if file name " + filename + " has _copy_ suffix.");
+    } else {
+      taskId = m.group(1);
+      copyFileSuffix = m.group(4);
+    }
+
+    LOG.debug("Filename: " + filename + " TaskId: " + taskId + " CopySuffix: " + copyFileSuffix);
+    if (taskId != null && copyFileSuffix != null) {
+      return true;
+    }
+
+    return false;
   }
 
   public static String getNameMessage(Exception e) {
@@ -2680,7 +2743,7 @@ public final class Utilities {
    * first time it is caught, or SQLTransientException when the maxRetries has reached.
    */
   public static <T> T executeWithRetry(SQLCommand<T> cmd, PreparedStatement stmt,
-      int baseWindow, int maxRetries)  throws SQLException {
+      long baseWindow, int maxRetries)  throws SQLException {
 
     Random r = new Random();
     T result = null;
@@ -2722,7 +2785,7 @@ public final class Utilities {
    * first time it is caught, or SQLTransientException when the maxRetries has reached.
    */
   public static Connection connectWithRetry(String connectionString,
-      int waitWindow, int maxRetries) throws SQLException {
+      long waitWindow, int maxRetries) throws SQLException {
 
     Random r = new Random();
 
@@ -2764,7 +2827,7 @@ public final class Utilities {
    * first time it is caught, or SQLTransientException when the maxRetries has reached.
    */
   public static PreparedStatement prepareWithRetry(Connection conn, String stmt,
-      int waitWindow, int maxRetries) throws SQLException {
+      long waitWindow, int maxRetries) throws SQLException {
 
     Random r = new Random();
 
@@ -2804,7 +2867,7 @@ public final class Utilities {
    * @param r a random generator.
    * @return number of milliseconds for the next wait time.
    */
-  public static long getRandomWaitTime(int baseWindow, int failures, Random r) {
+  public static long getRandomWaitTime(long baseWindow, int failures, Random r) {
     return (long) (
           baseWindow * failures +     // grace period for the last round of attempt
           baseWindow * (failures + 1) * r.nextDouble()); // expanding time window for each failure
@@ -3012,7 +3075,7 @@ public final class Utilities {
    * so we don't want to depend on scratch dir and context.
    */
   public static List<Path> getInputPathsTez(JobConf job, MapWork work) throws Exception {
-    String scratchDir = HiveConf.getVar(job, HiveConf.ConfVars.SCRATCHDIR);
+    String scratchDir = job.get(DagUtils.TEZ_TMP_DIR_KEY);
 
     // we usually don't want to create dummy files for tez, however the metadata only
     // optimization relies on it.
@@ -3313,7 +3376,7 @@ public final class Utilities {
   /**
    * Returns true if a plan is both configured for vectorized execution
    * and vectorization is allowed. The plan may be configured for vectorization
-   * but vectorization dissalowed eg. for FetchOperator execution.
+   * but vectorization disallowed eg. for FetchOperator execution.
    */
   public static boolean isVectorMode(Configuration conf) {
     if (HiveConf.getBoolVar(conf, HiveConf.ConfVars.HIVE_VECTORIZATION_ENABLED) &&
@@ -3462,13 +3525,13 @@ public final class Utilities {
     return createDirsWithPermission(conf, mkdir, fsPermission, recursive);
   }
 
-  private static void resetConfAndCloseFS (Configuration conf, boolean unsetUmask, 
+  private static void resetConfAndCloseFS (Configuration conf, boolean unsetUmask,
       String origUmask, FileSystem fs) throws IOException {
     if (unsetUmask) {
       if (origUmask != null) {
-        conf.set("fs.permissions.umask-mode", origUmask);
+        conf.set(FsPermission.UMASK_LABEL, origUmask);
       } else {
-        conf.unset("fs.permissions.umask-mode");
+        conf.unset(FsPermission.UMASK_LABEL);
       }
     }
 
@@ -3482,10 +3545,10 @@ public final class Utilities {
         recursive);
 
     if (recursive) {
-      origUmask = conf.get("fs.permissions.umask-mode");
+      origUmask = conf.get(FsPermission.UMASK_LABEL);
       // this umask is required because by default the hdfs mask is 022 resulting in
       // all parents getting the fsPermission & !(022) permission instead of fsPermission
-      conf.set("fs.permissions.umask-mode", "000");
+      conf.set(FsPermission.UMASK_LABEL, "000");
     }
 
     FileSystem fs = ShimLoader.getHadoopShims().getNonCachedFileSystem(mkdirPath.toUri(), conf);
